rename:: 코드 디렉토리 정리, 분리 - Timetable 내 컴포넌트가 많아짐에 따라, Component 내에서도 Page 별 Directory로 관리 - MyTimetablePage 내에서 관리되던 StatusBar 컴포넌트로 분리 (효율적인 Data Fetching을 위해 리팩토링 예정
